pip-ify and script-ify this code
